address rewinder review cleanup
