+++ electron/src/lib/time.ts
@@ -1,0 +1,15 @@
+export function formatDurationSeconds(seconds: number): string {
+  if (seconds < 60) {
+    return `${Math.ceil(seconds)} s`;
+  }
+
+  const totalMinutes = Math.ceil(seconds / 60);
+  const hours = Math.floor(totalMinutes / 60);
+  const minutes = totalMinutes % 60;
+
+  if (hours === 0) {
+    return `${totalMinutes} min`;
+  }
+
+  return minutes === 0 ? `${hours} h` : `${hours} h ${minutes} min`;
+}
